Reorganize to-do list structure in review section
