+++ cfgov/unprocessed/apps/youth-employment-success/js/store/index.js
@@ -1,9 +1,9 @@
 import YesStore from './yes-store';
-import budgetReducer from './reducers/budget-reducer';
-import { combineReducers } from './util';
-import goalReducer from './reducers/goal-reducer';
-import routeOptionReducer from './reducers/route-option-reducer';
-import choiceReducer from './reducers/choice-reducer';
+import budgetReducer from '../reducers/budget-reducer';
+import { combineReducers } from '../util';
+import goalReducer from '../reducers/goal-reducer';
+import routeOptionReducer from '../reducers/route-option-reducer';
+import choiceReducer from '../reducers/choice-reducer';
 
 /**
  * Function to create a new store instance
